requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.54.0
django-bootstrap-v5==1.0.11
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-rest-auth==0.9.5
django-taggit==4.0.0
djangorestframework==3.14.0
fontawesomefree==6.4.2
idna==3.4
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.8.1.78
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7